+++ Changelog.md
@@ -1,8 +1,11 @@
 # V0.5.1
 
-- add possibility to set default settings for a tool
+- add possibility to update default settings for a tool
+- fix png export not working with url pic
+- fix pictures missing id
 - fix text tool
 - double click for editing text is now working on mobile
+- add bold and italic for text
 
 # V0.5.0